User acceptance of clients is now optional
When registering a client in the client-registry, the option
requires-user-acceptance? flag can be used to indicate whether the
client is to be considered special, in that it does not require user
acceptance.

This is intended to support the registration of default or built-in
clients through the same registration mechanism. Of course,
implementations of the client-registry can ignore this option, treating
it merely as a hint, in determining which clients are to be deemed
built-in.