Refactoring (see below)
* session ids have been named to be unambiguous.

There are now just 2 session stores (one for the auth-server, one for
the client).

* :cylon/identity replaced with :cylon/subject-identifier

The rationale here is that :cylon/identity is ambiguous. OpenID uses the
term 'subject' for the user. A subject identifier is something that
uniquely identifies the subject, that is all. It can be a user-id, but
can equally be an email address. Cylon, as a library, should be agnostic.

* stub-emailer

A mod :stub-emailer has been added. Using it replaces the live sendgrid
emailer with a dummy one that just logs

e.g. (set-env! :ui :stub-emailer)

* src/cylon/core.bak deleted - it is no longer useful, even as a
  reference.

* cylon/impl/signup split and moved into cylon/signup, as per session
  and token-store

* secondary signup authorization server endpoint and authentication
  interaction has been removed. welcome page has also been removed

signup has been folded into a single component, incorporating an
optional totp notification. I don't think the extra
complexity of a parallel authentication interaction is necessary for
signup. It isn't necessary to split the interaction into multiple steps
for totp, since the totp is generated in the POST and displayed after
sign-up.

The key observation is that authentication interaction is merely a
sub-routine of the OAuth2 authorize-client process. The sign-up process
can indicate that the user is signed-in via the :cylon/authenticated?
flag in the authorization sesssion. The flag can be set or cleared
depending on the policy of forcing users to re-login post-signup (or
not). Note: The signup process does not subvert the usual OAuth2
authorize-client process which is still invoked by accessing a
restricted-handler. However, when the process invokes the authentication
'sub-routine', the outcome returns a value immediately for the signed-in
user.