Update co-dependency support
The relevant example of a benefit is that components that provide bidi routes can also access the router that combines them. This is essential for constructing routes between components with bidi's path-for. In the past, this was achieved ad-hoc, by some interception of the Ring request allowing the insertion into the request.

Such hacks are hard to generalize. Co-dependencies provide a generalization that can be used for other cases, including ones not involving Ring.

In the same way that component dependencies are declared with 'using', co-dependencies are declared with 'co-using'.

See https://github.com/juxt/modular.co-dependency for more details.

The impact of this change is that symbols previously named router are now named *router
The *-prefix to the symbol is an indication that the symbol is a reference (in this case, to a promise), so you must de-reference it first. You cannot use co-dependencies during the Lifecycle/start function, but can you them afterwards.

(Juan and I did attempt some clever machinery to hide the internal use of promises, but I feel that obscuring this fact is more damaging to overall understanding)